Fixed the problem(s)
The redirect problem was due to a change in the way I was shooting my request.
Instend of using the Session.get method, I was using the Session.send method.
This seems to have upset my code. Then had a bs4 html parsing errors and that
was because html5lib had not been reinstalled after moving to a virtual env.